requirements.txt
streamlit
langchain
langchain_openai
pandas
python-dotenv
openpyxl
datetime
langchain_community
requests